+++ prime_sdk/examples/allocations/create_portfolio_allocations.py
@@ -71,7 +71,7 @@ def main():
         order_ids=order_ids,
         allocation_legs=[allocation_leg],
         size_type=SizeType(args.size_type),
-        remainder_destination_portfolio_id=args.remainder_destination_portfolio_id,
+        remainder_destination_portfolio=args.remainder_destination_portfolio_id,
     )
 
     try: